Reduce Sonar noise in Maestro harness
