+++ scripts/maestro/prepare-android-scenario.mjs
@@ -1,5 +1,5 @@
-import { spawn } from 'node:child_process';
-import { defaultDeviceBaseUrl, exampleAppDir, getScenario } from './scenarios.mjs';
+import { runCommand } from './command.mjs';
+import { createBuildEnv, exampleAppDir, getScenario } from './scenarios.mjs';
 
 const scenarioId = process.argv[2];
 
@@ -9,36 +9,14 @@ if (!scenarioId) {
 
 const scenario = getScenario(scenarioId);
 
-function runCommand(command, args, options = {}) {
-  return new Promise((resolve, reject) => {
-    const child = spawn(command, args, {
-      cwd: options.cwd,
-      env: options.env,
-      stdio: 'inherit',
-    });
-
-    child.on('exit', (code) => {
-      if (code === 0) {
-        resolve();
-        return;
-      }
-
-      reject(new Error(`${command} ${args.join(' ')} failed with exit code ${code}`));
-    });
-  });
-}
-
 const env = {
-  ...process.env,
+  ...createBuildEnv({
+    scenarioId: scenario.id,
+    directUpdate: scenario.directUpdate,
+    appLabel: scenario.builtinLabel,
+  }),
   CAPGO_AUTO_UPDATE: 'true',
   CAPGO_DIRECT_UPDATE: scenario.directUpdate,
-  CAPGO_UPDATE_URL: `${defaultDeviceBaseUrl}/api/updates/${scenario.id}`,
-  CAPGO_STATS_URL: `${defaultDeviceBaseUrl}/api/stats`,
-  CAPGO_CHANNEL_URL: `${defaultDeviceBaseUrl}/api/channel`,
-  VITE_CAPGO_APP_LABEL: scenario.builtinLabel,
-  VITE_CAPGO_SCENARIO: scenario.id,
-  VITE_CAPGO_DIRECT_UPDATE: scenario.directUpdate,
-  VITE_CAPGO_SERVER_URL: `${defaultDeviceBaseUrl}/api/updates/${scenario.id}`,
 };
 
 await runCommand('bun', ['run', 'build'], {
